+++ CHANGES.md
@@ -1,7 +1,10 @@
-## Unreleased
+## 0.0.1~preview-0.1 (2024-08-19)
 
 ### Added
 
+- Added basic support for `readme`.
+- Added `Arg.named_multi`.
+- Added param helpers: `stringable`,`validated strings`, `comma_separated`.
 - Basic support for positional arguments.
 - Enabled instrumentation.
 - Adopted OCaml Code of Conduct.
@@ -10,10 +13,10 @@
 
 ### Changed
 
+- Internal changes to AST to make it more consistent.
+- Improve generation of man pages when using `cmdliner` as target.
 - Update tutorial to include positional arguments.
 
 ### Fixed
 
 - Translation to `core.command` requires `(unit -> _) Command.t`
-
-### Removed